requirements.txt
requests
telebot
matplotlib
numpy
tensorflow
scikit-image
pytelegrambotapi
pymongo
python-telegram-bot==20.0a4
beautifulsoup4